requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pymysql==1.1.1
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.32.3
PyMuPDF==1.24.14
